+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Changed
+
+## [1.7.1] - 2018-12-21
+### Added
 - Remove SendGrid, GADWP and Email Address Encored notifications from dashboard
+
+### Changed
 - Do not trust deactivation hook on air_helper_deactivated_without_version nag save, maybe do it in admin_init
 
 ## [1.7.0] - 2018-12-05